+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 1.2.0
 * Fix multiple naming typos in public APIs (Breaking Changes).
+* Add latest candle point marker and time chart indicator hiding options
+* Refactor ChartType from enum to sealed class with integrated style encapsulation (Breaking Changes).
 
 
 ## 1.1.1